+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 ### Fixed
 - `CrawlAction` element name bug.
 - Check if CrawlAction is unique only if it is not a link.
+- Replaced deprecated function `page.waitForTimeout`.
 
 ## [0.1.0] - 2024-03-30
 ### Added